# docs/system/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseIdResultDTO": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IdResultDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseObject": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseVoid": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateDepartmentDTO": {
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "parentId": {
            "format": "int64",
            "type": "integer"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "name",
          "parentId",
          "sortOrder",
          "version"
        ],
        "type": "object"
      },
...
      "DepartmentEnabledDTO": {
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "enabled",
          "version"
        ],
        "type": "object"
      },
...
      "IdResultDTO": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "InventoryLineDTO": {
        "properties": {
          "expectedVersion": {
            "format": "int32",
            "type": "integer"
          },
          "itemId": {
            "type": "string"
          },
          "lineRemark": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "quantity": {
            "maxLength": 32,
            "minLength": 0,
            "type": "string"
          },
          "targetLocationId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "itemId",
          "locationId",
          "quantity"
        ],
        "type": "object"
      },
      "InventoryRequestDTO": {
        "properties": {
          "correctedOperationId": {
            "format": "int64",
            "type": "integer"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InventoryLineDTO"
            },
            "maxItems": 100,
            "minItems": 0,
            "type": "array"
          },
          "remark": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "lines",
          "requestId"
        ],
        "type": "object"
      },
      "ItemCreateDTO": {
        "properties": {
          "baseUnit": {
            "maxLength": 32,
            "minLength": 0,
            "type": "string"
          },
          "code": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "baseUnit",
          "code",
          "name"
        ],
        "type": "object"
      },
      "ItemUpdateDTO": {
        "properties": {
          "baseUnit": {
            "maxLength": 32,
            "minLength": 0,
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "baseUnit",
          "name",
          "version"
        ],
        "type": "object"
      },
      "LocationCreateDTO": {
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "warehouseId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "warehouseId"
        ],
        "type": "object"
      },
      "LocationUpdateDTO": {
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "version"
        ],
        "type": "object"
      },
...
      "UpdateDepartmentDTO": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "parentId": {
            "format": "int64",
            "type": "integer"
          },
          "sortOrder": {
            "format": "int32",
            "type": "integer"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "parentId",
          "sortOrder",
          "version"
        ],
        "type": "object"
      },
...
      "WarehouseCreateDTO": {
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "departmentId": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "code",
          "departmentId",
          "name"
        ],
        "type": "object"
      },
      "WarehouseUpdateDTO": {
        "properties": {
          "departmentId": {
            "format": "int64",
            "type": "integer"
          },
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "departmentId",
          "name",
          "version"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "由 contract profile 在运行时生成的 0.1 API 契约。",
    "title": "Internal Admin Template API",
    "version": "0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/departments": {
      "post": {
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIdResultDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "department-controller"
        ]
      }
    },
...
    "/api/departments/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "department-controller"
        ]
      },
      "put": {
        "operationId": "update_2",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "department-controller"
        ]
      }
    },
    "/api/departments/{id}/enabled": {
      "put": {
        "operationId": "setEnabled",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentEnabledDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "department-controller"
        ]
      }
    },
...
    "/api/warehouse/inbound": {
      "post": {
        "operationId": "inbound",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
    "/api/warehouse/items": {
      "get": {
        "operationId": "items",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      },
      "post": {
        "operationId": "createItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIdResultDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
    "/api/warehouse/items/{id}": {
      "put": {
        "operationId": "updateItem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
    "/api/warehouse/locations": {
      "post": {
        "operationId": "createLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIdResultDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
...
    "/api/warehouse/locations/{id}": {
      "put": {
        "operationId": "updateLocation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
    "/api/warehouse/locations/{warehouseId}": {
      "get": {
        "operationId": "locations",
        "parameters": [
          {
            "in": "path",
            "name": "warehouseId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      }
    },
...
    "/api/warehouse/outbound": {
      "post": {
        "operationId": "outbound",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
...
    "/api/warehouse/stocktake": {
      "post": {
        "operationId": "stocktake",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
    "/api/warehouse/transfer": {
      "post": {
        "operationId": "transfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
    "/api/warehouse/warehouses": {
      "get": {
        "operationId": "warehouses",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      },
      "post": {
        "operationId": "createWarehouse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIdResultDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
...
    "/api/warehouse/warehouses/{id}": {
      "put": {
        "operationId": "updateWarehouse",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ],
...
}